[UI/#18] 옐로하기 잠금 화면 레이아웃 구현
