+++ app/src/main/java/com/yello/presentation/main/MainActivity.kt
@@ -12,7 +12,9 @@ import com.yello.presentation.main.myyello.MyYelloFragment
 import com.yello.presentation.main.profile.ProfileFragment
 import com.yello.presentation.main.recommend.RecommendFragment
 import com.yello.presentation.main.yello.YelloFragment
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : BindingActivity<ActivityMainBinding>(R.layout.activity_main) {
 
     override fun onCreate(savedInstanceState: Bundle?) {
