+++ app/src/main/java/com/yello/presentation/main/yello/lock/YelloLockFragment.kt
@@ -1,4 +1,4 @@
-package com.yello.presentation.onboarding.fragment
+package com.yello.presentation.main.yello.lock
 
 import android.os.Bundle
 import android.view.View
