+++ app/src/main/res/layout/fragment_yello.xml
@@ -7,7 +7,8 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layout_yello"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:background="@color/black">
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fcv_yello"
