+++ app/src/main/res/layout/fragment_yello_wait.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.onboarding.fragment.YelloWaitFragment">
+    tools:context=".presentation.main.yello.wait.YelloWaitFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layout_wait"
